requirements.txt
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
ffmpeg-python==0.2.0
future==0.18.2
idna==2.9
requests==2.23.0
urllib3==1.25.8